# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Figranium API",
    "description": "REST API for Figranium — a browser automation and web scraping platform (headless scraping, headful VNC browser sessions, and AI agent execution).",
    "version": "0.14.4"
  },
  "servers": [
    { "url": "http://localhost:11345", "description": "Local server" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key configured via /api/settings/api-key."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Machine-readable error code." },
          "message": { "type": "string" },
          "details": {},
          "detail": {}
        },
        "required": ["error"]
      },
      "SuccessFlag": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true }
        },
        "required": ["success"]
      },
...
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  },
  "security": [{ "sessionAuth": [] }, { "apiKeyAuth": [] }],
  "paths": {
...
    "/api/auth/check-setup": {
      "get": {
        "tags": ["Auth"],
        "summary": "Check whether initial setup (first user) has been completed",
        "security": [],
        "responses": {
          "200": {
            "description": "Setup status",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "setupRequired": { "type": "boolean" } } } } }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/auth/setup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create the first user account",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "email": { "type": "string", "maxLength": 255 },
                  "password": { "type": "string", "minLength": 8, "maxLength": 128 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessFlag" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "description": "Already set up", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged in", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessFlag" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  }
}